Deathmessage | Raidlog | FixYourName
Deathmessage - Fixed loop issue (missed an end) that would cause the
print function to print as many people there were in the server due to
the for loop not being closed.

raidlog - Logs to file with date & time, logs to admin chat. Uses
SteamID's instead of 64 ID's as they turn into e^16.

FixYourName - Removed Oxmin dependency and replaced it with Flags.